use hippie + adapt config for API testing

+++ test/root.coffee
@@ -1,11 +1,17 @@
 _ = require 'underscore'
 mongoose = require 'mongoose'
+hippie = require 'hippie'
+
 settings = require '../ethic/settings.js'
+server = require '../ethic/index.js'
 
 before ->
-  mongoose.connect 'mongodb://localhost/ethic-test'
+  @api = hippie server
+  if mongoose.connection.readyState == 0
+    mongoose.connect settings.mongoUri, settings.mongoOptions
 
 after ->
-  _.each mongoose.connection.collections, (col, name) ->
-    col.drop()
-  mongoose.disconnect()
+  if mongoose.connection.readyState > 0
+    _.each mongoose.connection.collections, (col, name) ->
+      col.drop()
+    mongoose.disconnect()
